게시물 첨부파일 파일 이

+++ src/main/java/egovframework/example/admin/boards/service/BoardsService.java
@@ -8,7 +8,7 @@ public interface BoardsService {
 	int countByBoard(String boardId) throws Exception;
 	List<Map<String, Object>> findByBoard(String boardId, int size, int offset) throws Exception;
 	PostVO getPostDetail(String postId) throws Exception;
-	int createPost(PostVO dto) throws Exception;
+	Long createPost(PostVO dto) throws Exception;
 	int updatePost(PostVO dto) throws Exception;
 	int deletePost(Map<String,Object> param) throws Exception;
 }
